chore: moved metric related types to ~/types/metrics.ts
