+++ src/server/api/routers/metrics/fetchMetricsForProjects.ts
@@ -1,10 +1,9 @@
-import {
-  type MetricWithProjects,
-  type OSOMetric,
-  fetchImpactMetrics,
-  mapMetrics,
-} from "~/utils/fetchMetrics";
-import type { ImpactMetricsQuery } from "~/utils/fetchMetrics/types";
+import { fetchImpactMetrics, mapMetrics } from "~/utils/fetchMetrics";
+import type {
+  ImpactMetricsQuery,
+  MetricWithProjects,
+} from "~/utils/fetchMetrics/types";
+import type { OSOMetric } from "~/types/metrics";
 import { mockedApprovedProjects } from "./mocks";
 
 export async function fetchMetricsForProjects({
