+++ src/utils/fetchMetrics/index.ts
@@ -1,9 +1,2 @@
-export type {
-  OSOMetric,
-  OSOMetrics,
-  MetricProject,
-  MetricWithProjects,
-} from "./types";
-
 export { fetchImpactMetrics } from "./fetchImpactMetrics";
 export { mapMetrics } from "./mapMetrics";
